requirements.txt
langchain==0.2.9
langchain_community==0.2.7
langchain_google_genai==1.0.7
python-dotenv==1.0.1
